requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx  # For async HTTP requests to Ollama
pydantic # FastAPI dependency, good to include explicitly
protobuf # For SingularityNET .proto compilation